refactored particle loop such that the implemetentations for each container are with the container
