+++ include/cell_dat.hpp
@@ -12,6 +12,8 @@
 
 #include "access.hpp"
 #include "compute_target.hpp"
+#include "loop/access_descriptors.hpp"
+#include "loop/particle_loop_base.hpp"
 #include "profiling.hpp"
 #include "typedefs.hpp"
 
@@ -119,6 +121,119 @@ template <typename T> struct CellDatConstDeviceTypeConst {
   int nrow;
 };
 
+class ParticleGroup;
+template <typename T> class CellDatConst;
+
+/**
+ *  Defines the access implementations and types for CellDatConst objects.
+ */
+namespace Access::CellDatConst {
+
+/**
+ * Access:CellDatConst::Read<T> and Access:CellDatConst::Add<T> are the
+ * kernel argument types for accessing CellDatConst data in a kernel.
+ */
+template <typename T> struct Read {
+  /// Pointer to underlying data for the array.
+  Read() = default;
+  T const *ptr;
+  int nrow;
+  inline const T at(const int row, const int col) {
+    return ptr[nrow * col + row];
+  }
+  inline const T &operator[](const int component) { return ptr[component]; }
+};
+
+/**
+ * Access:CellDatConst::Read<T> and Access:CellDatConst::Add<T> are the
+ * kernel argument types for accessing CellDatConst data in a kernel.
+ */
+template <typename T> struct Add {
+  /// Pointer to underlying data for the array.
+  Add() = default;
+  T *ptr;
+  int nrow;
+  inline T fetch_add(const int row, const int col, const T value) {
+    sycl::atomic_ref<T, sycl::memory_order::relaxed, sycl::memory_scope::device>
+        element_atomic(ptr[nrow * col + row]);
+    return element_atomic.fetch_add(value);
+  }
+};
+
+} // namespace Access::CellDatConst
+
+namespace ParticleLoopImplementation {
+
+/**
+ *  Loop parameter for read access of a CellDatConst.
+ */
+template <typename T> struct LoopParameter<Access::Read<CellDatConst<T>>> {
+  using type = CellDatConstDeviceTypeConst<T>;
+};
+/**
+ *  Loop parameter for add access of a CellDatConst.
+ */
+template <typename T> struct LoopParameter<Access::Add<CellDatConst<T>>> {
+  using type = CellDatConstDeviceType<T>;
+};
+
+/**
+ *  KernelParameter type for read access to a CellDatConst.
+ */
+template <typename T> struct KernelParameter<Access::Read<CellDatConst<T>>> {
+  using type = Access::CellDatConst::Read<T>;
+};
+/**
+ *  KernelParameter type for add access to a CellDatConst.
+ */
+template <typename T> struct KernelParameter<Access::Add<CellDatConst<T>>> {
+  using type = Access::CellDatConst::Add<T>;
+};
+
+/**
+ *  Function to create the kernel argument for CellDatConst read access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx,
+                              CellDatConstDeviceTypeConst<T> &rhs,
+                              Access::CellDatConst::Read<T> &lhs) {
+  T const *ptr = rhs.ptr + cellx * rhs.stride;
+  lhs.ptr = ptr;
+  lhs.nrow = rhs.nrow;
+}
+/**
+ *  Function to create the kernel argument for CellDatConst add access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx,
+                              CellDatConstDeviceType<T> &rhs,
+                              Access::CellDatConst::Add<T> &lhs) {
+  T *ptr = rhs.ptr + cellx * rhs.stride;
+  lhs.ptr = ptr;
+  lhs.nrow = rhs.nrow;
+}
+
+/**
+ * Method to compute access to a CellDatConst (read)
+ */
+template <typename T>
+inline CellDatConstDeviceTypeConst<T>
+create_loop_arg(ParticleGroup *particle_group, sycl::handler &cgh,
+                Access::Read<CellDatConst<T> *> &a) {
+  return a.obj->impl_get_const();
+}
+/**
+ * Method to compute access to a CellDatConst (add)
+ */
+template <typename T>
+inline CellDatConstDeviceType<T>
+create_loop_arg(ParticleGroup *particle_group, sycl::handler &cgh,
+                Access::Add<CellDatConst<T> *> &a) {
+  return a.obj->impl_get();
+}
+
+} // namespace ParticleLoopImplementation
+
 /**
  *  Container that allocates on the device a matrix of fixed size nrow X ncol
  *  for N cells. Data stored in column major format. i.e. Data order from
@@ -126,6 +241,14 @@ template <typename T> struct CellDatConstDeviceTypeConst {
  */
 template <typename T> class CellDatConst {
   template <typename KERNEL, typename... ARGS> friend class ParticleLoop;
+  friend CellDatConstDeviceTypeConst<T>
+  ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Read<CellDatConst<T> *> &a);
+  friend CellDatConstDeviceType<T>
+  ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Add<CellDatConst<T> *> &a);
 
 private:
   T *d_ptr;
