+++ include/containers/local_array.hpp
@@ -2,6 +2,8 @@
 #define _NESO_PARTICLES_LOCAL_ARRAY_H_
 
 #include "../compute_target.hpp"
+#include "../loop/access_descriptors.hpp"
+#include "../loop/particle_loop_base.hpp"
 #include <memory>
 #include <optional>
 #include <vector>
@@ -10,6 +12,157 @@ namespace NESO::Particles {
 // Forward declaration of ParticleLoop such that LocalArray can define
 // ParticleLoop as a friend class.
 template <typename KERNEL, typename... ARGS> class ParticleLoop;
+class ParticleGroup;
+template <typename T> class LocalArray;
+
+template <typename T> using LocalArrayImplGetT = T *;
+template <typename T> using LocalArrayImplGetConstT = T const *;
+
+/**
+ *  Defines the access implementations and types for LocalArray objects.
+ */
+namespace Access::LocalArray {
+
+/**
+ * Access:LocalArray::Read<T>, Access::LocalArray::Write<T> and
+ * Access:LocalArray::Add<T> are the kernel argument types for accessing
+ * LocalArray data in a kernel.
+ */
+/**
+ * ParticleLoop access type for LocalArray Read access.
+ */
+template <typename T> struct Read {
+  /// Pointer to underlying data for the array.
+  Read() = default;
+  T const *ptr;
+  const T at(const int component) { return ptr[component]; }
+  const T &operator[](const int component) { return ptr[component]; }
+};
+
+/**
+ * ParticleLoop access type for LocalArray Add access.
+ */
+template <typename T> struct Add {
+  /// Pointer to underlying data for the array.
+  Add() = default;
+  T *ptr;
+  /**
+   * The local array is local to the MPI rank where the partial sum is a
+   * meaningful value.
+   */
+  inline T fetch_add(const int component, const T value) {
+    sycl::atomic_ref<T, sycl::memory_order::relaxed, sycl::memory_scope::device>
+        element_atomic(ptr[component]);
+    return element_atomic.fetch_add(value);
+  }
+};
+
+/**
+ * ParticleLoop access type for LocalArray Write access.
+ */
+template <typename T> struct Write {
+  /// Pointer to underlying data for the array.
+  Write() = default;
+  T *ptr;
+  T at(const int component) { return ptr[component]; }
+  T &operator[](const int component) { return ptr[component]; }
+};
+
+} // namespace Access::LocalArray
+
+namespace ParticleLoopImplementation {
+
+/**
+ *  Loop parameter for read access of a LocalArray.
+ */
+template <typename T> struct LoopParameter<Access::Read<LocalArray<T>>> {
+  using type = T const *;
+};
+/**
+ *  Loop parameter for write access of a LocalArray.
+ */
+template <typename T> struct LoopParameter<Access::Write<LocalArray<T>>> {
+  using type = T *;
+};
+/**
+ *  Loop parameter for add access of a LocalArray.
+ */
+template <typename T> struct LoopParameter<Access::Add<LocalArray<T>>> {
+  using type = T *;
+};
+
+/**
+ *  KernelParameter type for read access to a LocalArray.
+ */
+template <typename T> struct KernelParameter<Access::Read<LocalArray<T>>> {
+  using type = Access::LocalArray::Read<T>;
+};
+/**
+ *  KernelParameter type for write access to a LocalArray.
+ */
+template <typename T> struct KernelParameter<Access::Write<LocalArray<T>>> {
+  using type = Access::LocalArray::Write<T>;
+};
+/**
+ *  KernelParameter type for add access to a LocalArray.
+ */
+template <typename T> struct KernelParameter<Access::Add<LocalArray<T>>> {
+  using type = Access::LocalArray::Add<T>;
+};
+/**
+ *  Function to create the kernel argument for LocalArray read access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx, T const *rhs,
+                              Access::LocalArray::Read<T> &lhs) {
+  lhs.ptr = rhs;
+}
+/**
+ *  Function to create the kernel argument for LocalArray write access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx, T *rhs,
+                              Access::LocalArray::Write<T> &lhs) {
+  lhs.ptr = rhs;
+}
+/**
+ *  Function to create the kernel argument for LocalArray add access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx, T *rhs,
+                              Access::LocalArray::Add<T> &lhs) {
+  lhs.ptr = rhs;
+}
+
+/**
+ * Method to compute access to a LocalArray (read)
+ */
+template <typename T>
+inline LocalArrayImplGetConstT<T>
+create_loop_arg(ParticleGroup *particle_group, sycl::handler &cgh,
+                Access::Read<LocalArray<T> *> &a) {
+  return a.obj->impl_get_const();
+}
+/**
+ * Method to compute access to a LocalArray (write)
+ */
+template <typename T>
+inline LocalArrayImplGetT<T>
+create_loop_arg(ParticleGroup *particle_group, sycl::handler &cgh,
+                Access::Write<LocalArray<T> *> &a) {
+  return a.obj->impl_get();
+}
+/**
+ * Method to compute access to a LocalArray (add)
+ */
+template <typename T>
+inline LocalArrayImplGetT<T> create_loop_arg(ParticleGroup *particle_group,
+                                             sycl::handler &cgh,
+                                             Access::Add<LocalArray<T> *> &a) {
+  return a.obj->impl_get();
+}
+
+} // namespace ParticleLoopImplementation
 
 /**
  * Container to hold an array of values on each MPI rank.
@@ -17,6 +170,16 @@ template <typename KERNEL, typename... ARGS> class ParticleLoop;
 template <typename T> class LocalArray {
   // This allows the ParticleLoop to access the implementation methods.
   template <typename KERNEL, typename... ARGS> friend class ParticleLoop;
+  friend LocalArrayImplGetConstT<T>
+  ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Read<LocalArray<T> *> &a);
+  friend LocalArrayImplGetT<T> ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Write<LocalArray<T> *> &a);
+  friend LocalArrayImplGetT<T> ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Add<LocalArray<T> *> &a);
 
 protected:
   std::shared_ptr<BufferDevice<T>> buffer;
@@ -25,13 +188,15 @@ protected:
    * Non-const pointer to underlying device data. Intended for friend access
    * from ParticleLoop.
    */
-  inline T *impl_get() { return this->buffer->ptr; }
+  inline LocalArrayImplGetT<T> impl_get() { return this->buffer->ptr; }
 
   /**
    * Const pointer to underlying device data. Intended for friend access
    * from ParticleLoop.
    */
-  inline T const *impl_get_const() { return this->buffer->ptr; }
+  inline LocalArrayImplGetConstT<T> impl_get_const() {
+    return this->buffer->ptr;
+  }
 
 public:
   /// The SYCLTarget the LocalArray is created on.
