+++ include/containers/sym_vector.hpp
@@ -1,6 +1,8 @@
 #ifndef _NESO_PARTICLES_SYM_VECTOR_H_
 #define _NESO_PARTICLES_SYM_VECTOR_H_
 #include "../compute_target.hpp"
+#include "../loop/access_descriptors.hpp"
+#include "../loop/particle_loop_base.hpp"
 #include "../particle_group.hpp"
 #include "../particle_spec.hpp"
 #include <initializer_list>
@@ -12,6 +14,112 @@ namespace NESO::Particles {
 // Forward declaration of ParticleLoop such that SymVector can define
 // ParticleLoop as a friend class.
 template <typename KERNEL, typename... ARGS> class ParticleLoop;
+template <typename T> using SymVectorImplGetT = T ****;
+template <typename T> using SymVectorImplGetConstT = T *const *const **;
+
+/**
+ *  Defines the access implementations and types for ParticleDat objects.
+ */
+namespace Access::SymVector {
+
+/**
+ * Access:SymVector::Read<T> and Access:SymVector::Read<T> are the
+ * kernel argument types for accessing particle data in a kernel via a
+ * SymVector.
+ */
+template <typename T> struct Read {
+  /// Pointer to underlying data.
+  T *const *const **ptr;
+  T const &at(const int dat_index, const int cell, const int layer,
+              const int component) {
+    return ptr[dat_index][cell][component][layer];
+  }
+  T const &at(const int dat_index,
+              const Access::LoopIndex::Read &particle_index,
+              const int component) {
+    return ptr[dat_index][particle_index.cell][component][particle_index.layer];
+  }
+};
+
+template <typename T> struct Write {
+  /// Pointer to underlying data.
+  T ****ptr;
+  T &at(const int dat_index, const int cell, const int layer,
+        const int component) {
+    return ptr[dat_index][cell][component][layer];
+  }
+  T &at(const int dat_index, const Access::LoopIndex::Read &particle_index,
+        const int component) {
+    return ptr[dat_index][particle_index.cell][component][particle_index.layer];
+  }
+};
+
+} // namespace Access::SymVector
+
+namespace ParticleLoopImplementation {
+/**
+ *  Loop parameter for read access of a SymVector.
+ */
+template <typename T> struct LoopParameter<Access::Read<SymVector<T>>> {
+  using type = T *const *const **;
+};
+/**
+ *  Loop parameter for write access of a SymVector.
+ */
+template <typename T> struct LoopParameter<Access::Write<SymVector<T>>> {
+  using type = T ****;
+};
+/**
+ *  KernelParameter type for read-only access to a SymVector.
+ */
+template <typename T> struct KernelParameter<Access::Read<SymVector<T>>> {
+  using type = Access::SymVector::Read<T>;
+};
+/**
+ *  KernelParameter type for write access to a SymVector.
+ */
+template <typename T> struct KernelParameter<Access::Write<SymVector<T>>> {
+  using type = Access::SymVector::Write<T>;
+};
+
+/**
+ * Method to compute access to a SymVector (read).
+ */
+template <typename T>
+inline SymVectorImplGetConstT<T>
+create_loop_arg(ParticleGroup *particle_group, sycl::handler &cgh,
+                Access::Read<SymVector<T> *> &a) {
+  return a.obj->impl_get_const();
+}
+/**
+ * Method to compute access to a SymVector (write).
+ */
+template <typename T>
+inline SymVectorImplGetT<T> create_loop_arg(ParticleGroup *particle_group,
+                                            sycl::handler &cgh,
+                                            Access::Write<SymVector<T> *> &a) {
+  return a.obj->impl_get();
+}
+
+/**
+ *  Function to create the kernel argument for SymVector read access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx,
+                              T *const *const **rhs,
+                              Access::SymVector::Read<T> &lhs) {
+  lhs.ptr = rhs;
+}
+/**
+ *  Function to create the kernel argument for SymVector write access.
+ */
+template <typename T>
+inline void create_kernel_arg(const int cellx, const int layerx, T ****rhs,
+                              Access::SymVector::Write<T> &lhs) {
+  lhs.ptr = rhs;
+}
+
+} // namespace ParticleLoopImplementation
 
 /**
  * Enables ParticleDats to be accessed in ParticleLoops with a number of
@@ -20,6 +128,13 @@ template <typename KERNEL, typename... ARGS> class ParticleLoop;
 template <typename T> class SymVector {
   // This allows the ParticleLoop to access the implementation methods.
   template <typename KERNEL, typename... ARGS> friend class ParticleLoop;
+  friend SymVectorImplGetConstT<T>
+  ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Read<SymVector<T> *> &a);
+  friend SymVectorImplGetT<T> ParticleLoopImplementation::create_loop_arg<T>(
+      ParticleGroup *particle_group, sycl::handler &cgh,
+      Access::Write<SymVector<T> *> &a);
 
 protected:
   std::vector<ParticleDatSharedPtr<T>> dats;
@@ -41,14 +156,14 @@ protected:
     this->dh_root_const_ptrs->host_to_device();
   }
 
-  inline T ****impl_get() {
+  inline SymVectorImplGetT<T> impl_get() {
     const int num_dats = this->dats.size();
     for (int dx = 0; dx < num_dats; dx++) {
       this->dats.at(dx)->impl_get();
     }
     return this->dh_root_ptrs->d_buffer.ptr;
   }
-  inline T *const *const **impl_get_const() {
+  inline SymVectorImplGetConstT<T> impl_get_const() {
     const int num_dats = this->dats.size();
     for (int dx = 0; dx < num_dats; dx++) {
       this->dats.at(dx)->impl_get_const();
